Fixed: occasionally-failing stopwatch unit test.
Unit test approach is to start a stopwatch, sleep, then check for the
elapsed time. However, our check for user-defined duration is looking for
precision in microseconds. At this granularity, we're bumping hard against
the inherent fuzziness in OS/thread scheduling. So we can't reliably do a
normal equals? check against elapsed time.

Since our test is more concerned with checking that we're getting the
correct granularity, *not* testing std::chrono that we're getting a
specific value, I changed the test condition to be more of an
order-of-magnitude check, rather than a paricular value.